changes to support NPM
