--- firebase.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "dreembase",
-  "private_key_id": "5d69d13612aa7ac75bd1e8a405a945f09b7b92cf",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIJQgIBADANBgkqhkiG9w0BAQEFAASCCSwwggkoAgEAAoICAQDQUm9ZEFIiNu1y\nNi5RIuAkf4ABiC20zUnG5qDtkMgcPTr4H+I1MTU4pXKmv2KMPl2KsLrQoKf/r3ul\nyqfGpeHQZgGQ/bpot78EBWHUP3P5HgsHHqofz/PY5JkION7FZa4v+6UuW/faxpWA\nSSQg9cLYsMmFP0qH5gIiWVH6pkkvKDZpcwID7DquxDfqpBgytZtzBYcT1C79Y5Xa\ntXjZZtR8r4XxHxwBUS8vDbF8pXdnRehpOUTe3O8l+B7bcyraqvSXfBdcFmYtpSMe\ndxzLX7gmKgEjYaJuUoNEyLy5pLNbTnTSPuq/Vijy3wJyGpACsIGrXLXzx1Sn/Wkw\nHrPM7aOEA/ScAEVIdbSuxckiJjRseAWY8wtgLI1v6l9aFZFoCT9voIfGtwKdcuMq\nxNjAOQeofNWbE66XbVNUVpbM5Jk63wLsIUJGOKeLM2uIMaovlRfnfB1CdXMPO3y8\nb2e1vqgZDWwdRSiQ7nB9GEUP3Osx0ohB7Bg2LmyT3pnIJiM9lCviBdQiB+f/uHMl\n7mCIcAUwztf8rvRrPT7CxyZQnbLfwksc+Uje/CixGT9TMO55z9JpiFvDitFhsB83\nnVa0uBBdUHVELS2oBzbIjStNCYJQUYhNcMf+myUxrxCHOap6P5jHhwBLop6dlDf3\nofCqZkFK/yUAxsqOFbuX+pVmY+7XJwIDAQABAoICADuJc702huK0ZAQiK/aQSxLj\nRYgqjTWA/jbULuX2XNqFPVcwdYvSNm2sCypQxC9km4NJovqBD8JpOqNVWcHSIslW\nfdgcS6wl53eUujLYt6JYzboJL4HMAtrX6kyJRFGSD8gy2TB4pEtoOc3iGWs6Eaew\nTsfAIOLZ/34lnOJPp9bMXVChx8+rMdJEzd5Ha4sCkBy9Va97ErrkPS22c4VeeBfy\nagtRYsWVmDCQb0s6Vvey6Qs5jM9MwSRO1zLMpVC9Tw4ELwDofedJrNHGJPqSrqRp\nmM2t16hD+LQt0pIT2hacQ5La3Gg2jZlr1t0ZblL2cviwlGfJbg5cxDU2cRh+Sh6q\nGUU/1izwGk4X39Yah3rf+vTXKQ3blRjhUWIU3H03bdbbMOjDxXnsywbz5iPTH2JH\nzSJ6Wla7NHlpjGqnm7PzIw+yivAew/PcVs0dbCFbNqzUFwt+T4SLEk6KhECXO9rP\n/iDySwOEk3QqY0jTKXkBehrqRx1G68rJrlkBfYJVye9B86WEkWtMccSRg7+Q9UZW\nNlkusMIhFZPGiRXpXmuGFzjW+6rCzBkEH1ESZ5F4fVm69rPAo5TfXHMgjdxJTexm\nBHXX5n5RWZtwhzaodAZYDMf6K7cU7xih2TyraFqtldw96wfD5sK40KJLhEgH/bUe\nvsZWsKLSyYdk5wPIGvxRAoIBAQD9XpJ5Bn3qQ/+B7LsqbQrFwLSJxEIsVd11AenC\nr+Kx0a+4EL5NQdW7KlVIuxDluDImHpjJsZ/mynP5sk7tEVlkRMXeTsd9woxZc81s\njgcXkrWyWBIX3GeearZ3184dC9HJOTHMfavcQBIHDzZdMSkusQuQ+bIa2PXtdTwG\n0gD79QTnGGcpEYHso7DSUBtJnGQR/jDE/nEDLZecktMsSh1W+97bS7cogxoTbhRb\nD5WgVfqXT6daY7Tx2vJwPRh4iRhwUweUUzHzFbuo3e2avVMhhhnQKzrQVAi69SrC\nJcz8tddjxZmaa7PCEQmejRvdZVFwqLuwQGhibGDqh4CkXmBrAoIBAQDSfCG7kuuX\nA3mvNa5vMkXZOZdgliGSuVs1A0sdyXbayk4KT1eDgAaWHP7wzDNXF/OMRhSxam6b\nSx3JQQxi36mC9QBFKBuBTAQDkFNbAKzN/xyXqlnQxJE3q5LtN+VJlUCgOcusjS7w\nJItaXlg/wb17GFkmMlcpyTn7Z8+53fbC2sFLMkPt6hfO23+ryqaLrVceYCkPaWfK\nfHoxUPnQlojlyoq1liFegEdx9A4SdydSTPmxpk09GBDOim9CdpFoQYMXuRrjF8XD\nw0cWglxZKjPisACiELvNI11e6ORmLGGDuD3pO4wTwTcT503Zv6FIg6o4yf+Gk2u3\nIbLOXFeIMuM1AoIBAQDHECCAWatf/SCn0XiTsqw+BlNR0hIzhxrBLUmdnla7YoVu\nVilNxaXIsU/kHqT99yPfczaTcKyyxOktivwbof/bmXTkXny5v7pJJYSF/2hvCuMa\nRbuhZ5c5jZniunjDtV+Tew43HZ2DJAuF5FvQCCnU92Cg/K0EDuXZNLx+l0wSqMS7\nVB7o00qCw3DPLATkZqzwvgdPHqFn39A4QdsL9wbqKmVDYk960tYP4fK4QJxMJoEF\nlJ1llOS7CExmBnpNaVj8Fx/jVyHtXGveFs7xPkIsGpH5TZ2dxAg/6xM3Cj/tZGjf\niIZQHsJSfNRT74WvAkVkAA3GQjRA7+mNGfzmJwgVAoIBAEeoyzM5CcwX+s2T0Rc6\nXl1eJhNUCYoSYOLjhCdX94b1aSxJZtn2NEfXB1SjsLbuCJSxJ0pfmdh08mcaJkVj\nMCrNjpH83aBfaN5tvHM2/54CRJG+BWX2BhqV5yer5/vtPFtQUe8lAd3wEyFpDVJn\ne0L1PySp7tLjF0JoTWhSETi1kSkwq7/L+t8KC6VJEPpswUQuKMx+bj+ZsWAo2WQS\nW/kfw6meuy2ZLrm1BTV0K4bxffDlVVZ/YvluWAKgq1B7PU5OTsERztqfGQyoxRf2\nLkP298A1QjrM13UAhQHkCi72At3sz/vJxB6amP6RJy+w47sBymTeu8VVS7LF7WcJ\nLCECggEAfmYWqDsn9CzDRiH3spVjlv5GmP3ilg83nXg6SSJV4x4txqQwqdeKxRL4\nvwPOOK0yhffbe31PGpra9nNN3eUKnfgLs6hcysAII7yF16VEXrbqPZNlPeEf8AZv\nymOjRzfjhlUYJ9D3Uk3+BkF31z12greWYxSVei77OLmWYmBblpwAQxiByuildfTx\nq4rwWiXDRkDaU+BQgyhRxEs/RVaF8dexQ70VJRA1USW2JMx5Z/NYVaW9D0p4jiwY\n5qka2xS9Z5ZBGvcjkNe1cLullO38LuttCmW0dJHF5FbYAa4D319ydrChjxUelszR\naVNrxYl8FSd503nQ6vZFs/m8GMd2xQ==\n-----END PRIVATE KEY-----\n",
-  "client_email": "dreembase@dreembase.iam.gserviceaccount.com",
-  "client_id": "104362130042575267591",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://accounts.google.com/o/oauth2/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/dreembase%40dreembase.iam.gserviceaccount.com"
-}
