+++ system/adapters/expressadapter.js
@@ -1,58 +1,5 @@
-// Proccess args
-var args = {}
-var argv = process.argv
-for(var lastkey = '', arg, i = 0; i<argv.length; i++){
-	arg = argv[i]
-	if(arg.charAt(0) == '-') lastkey = arg, args[lastkey] = true
-	else {
-		if(lastkey in args && args[lastkey] !== true){
-			if(!Array.isArray(args[lastkey])) args[lastkey] = [args[lastkey]]
-			args[lastkey].push(arg)
-		}
-		else args[lastkey] = arg
-	}
-}
-
 // Launch the platform setup
-require = require(__dirname + '/../base/define')
-
-if (args['-writefile']) {
-	Object.defineProperty(define, "$writefile", { value: true, writable: false });
-} else{
-	Object.defineProperty(define, "$writefile", { value: false, writable: false });
-}
-
-if (args['-unsafeorigin']) {
-	Object.defineProperty(define, "$unsafeorigin", { value: true, writable: false });
-} else{
-	Object.defineProperty(define, "$unsafeorigin", { value: false, writable: false });
-}
-
-define.$platform = 'nodejs'
-
-define.paths = {
-	'system':'$root/system',
-	'resources':'$root/resources',
-	'3d':'$root/classes/3d',
-	'behaviors':'$root/classes/behaviors',
-	'server':'$root/classes/server',
-	'ui':'$root/classes/ui',
-	'flow':'$root/classes/flow',
-	'testing':'$root/classes/testing',
-	'widgets':'$root/classes/widgets',
-	'sensors':'$root/classes/sensors',
-	'iot':'$root/classes/iot',
-	'examples':'$root/examples',
-	'apps':'$root/apps',
-	'docs':'$root/docs',
-	'test':'$root/test'
-}
-
-for (var key in define.paths) {
-	define['$' + key] = define.paths[key]
-}
-
-require('$system/base/math')
+require = require(__dirname + '/../../dreemgl')
 
 // Serves all the static files in that DreemGL will likely ask for (from define.paths)
 exports.initStatic = function(express, app) {
@@ -70,7 +17,7 @@ exports.requestHandler = function (req, res) {
 
 	var compositionserver = CompositionServer.compservers[compname]
 	if (!compositionserver) {
-		CompositionServer.compservers[compname] = compositionserver = new CompositionServer(args, compname, define.$compositionOptions)
+		CompositionServer.compservers[compname] = compositionserver = new CompositionServer(define.$args, compname, define.$compositionOptions)
 	}
 
 	compositionserver.request(req, res)
